Vuex, Exports, SASS fixes, LS

+++ src/components/tavern.vue
@@ -7,38 +7,37 @@ import Spinner from './common/Spinner';
 import { readLocalStorage,
          updateLocalStorage,
          clearLocalStorage } from '../helpers';
+import { mapGetters, 
+         mapMutations } from 'vuex';
+
 export default {
   name:'tavern',
   components: { Spinner },
   data() {
     return {
-      Character: [],
+      tavernList: [],
       isLoading: true,
       rollQueue: [],
-      // Local storage
-      enableLocalStorage: false,
       localCharactersList: [],
+      noLocalCharactersFound: false,
     }
   },
   // on Mounted
   mounted() {
-    if (this.enableLocalStorage && !readLocalStorage('localCharactersList')) {
-      window.location = '/create';
-    }
-
-    if (this.enableLocalStorage && readLocalStorage('localCharactersList')) {
+    // clearLocalStorage('localCharactersList');
+    // Get charsheets from LS by default
+    if (this.$store.state.enableLocalStorage) {
       this.isLoading = false;
-      // 1. Get list of locally stored characters
-      this.localCharactersList = readLocalStorage('localCharactersList');
-      console.log(this.localCharactersList);
-      // 2. 
-      for (let item of this.localCharactersList) {
-        this.Character.push(readLocalStorage(item));
+      if (!readLocalStorage('localCharactersList')) {
+        this.noLocalCharactersFound = true;
+      }
+      if (readLocalStorage('localCharactersList')) {
+        this.updateTavern();
       }
     }
 
-    if (!this.enableLocalStorage) {
-      // Get charsheets from DB
+    // Get charsheets from DB if LS is disabled
+    if (!this.$store.state.enableLocalStorage) {
       axios.get('https://dnd-charsheet-api.herokuapp.com/charsheets', {
         timeout: 5000
       })
@@ -46,7 +45,7 @@ export default {
           this.isLoading = false;
           console.log('Response.data came in:');
           console.log(response.data);
-          this.Character = response.data;
+          this.tavernList = response.data;
           // TODO: Check if there's local data
           // TODO: compare versions of local and remote data, when they are available
           this.updateRollQueue('Loading complete.');
@@ -59,19 +58,38 @@ export default {
   },
   // Methods
   methods: {
+    updateTavern () {
+      this.localCharactersList = readLocalStorage('localCharactersList');
+      this.tavernList = [];
+      for (let item of this.localCharactersList) {
+        // Fill the tavern with local characters one by one
+        this.tavernList.push(readLocalStorage(item));
+      }
+    },
+
     deleteCharacter (charID) {
-      console.log('Trying to delete: ' + charID);
-      this.isLoading = true;
-      axios.delete('https://dnd-charsheet-api.herokuapp.com/charsheets/delete/' + charID)
-      this.isLoading = false;
-      // .then((response) => {
-      //   console.log('Success deleting.');
-      //   this.isLoading = false;
-      // })
-      // .catch((error) => {
-      //   console.log('Failed to delete. ' + error);
-      //   this.isLoading = false;
-      // });
+      if (this.$store.state.enableLocalStorage) {
+        clearLocalStorage(charID);
+        let deleteFrom = this.localCharactersList.indexOf(charID);
+        this.localCharactersList.splice(deleteFrom, 1);
+        // Update locally stored list of names and character record
+        updateLocalStorage(this.localCharactersList, 'localCharactersList');
+        this.updateTavern();
+      }
+      if (!this.$store.state.enableLocalStorage) {
+        console.log('Trying to delete: ' + charID);
+        this.isLoading = true;
+        axios.delete('https://dnd-charsheet-api.herokuapp.com/charsheets/delete/' + charID)
+        this.isLoading = false;
+        // .then((response) => {
+        //   console.log('Success deleting.');
+        //   this.isLoading = false;
+        // })
+        // .catch((error) => {
+        //   console.log('Failed to delete. ' + error);
+        //   this.isLoading = false;
+        // });
+      }
     },
 
     updateRollQueue (string, note) {
